refactor: Extract api functions to own file, and reduce repetitive code
